refactor(tests): rearrange hub fixtures helper and update tool
