+++ tests/fixtures/update_services_fixtures
@@ -3,10 +3,12 @@
 from os import path
 import sys
 
-from tests.test_helpers import StoryscriptHubFixture
 
+sys.path.insert(
+    0, path.dirname(path.dirname(path.dirname(path.realpath(__file__)))))
 
-sys.path.insert(0, path.dirname(path.dirname(path.realpath(__file__))))
+from tests.e2e.helpers.StoryscriptHubFixture import StoryscriptHubFixture
+
 
 services = [
     'Arinono/microservice-validate',
